Implement virtual tables and method lookups
There is a bug with superclass constructors caused by an
inconsistency somewhere with the number of arguments. The subclass
constructor gets called with (self, args...), but the superclass
seems to get called as just (args...) where self is used as that
first argument. So if the subclass is init(n) : supser(n), then
looking up `n` accidently gets self.